// src/thunder-sdk/src/types/asset_manager/AssetManager.ts
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;

export type AddressInput = { bits: string };
export type AddressOutput = AddressInput;
export type AssetIdInput = { bits: string };
export type AssetIdOutput = AssetIdInput;
export type ContractIdInput = { bits: string };
export type ContractIdOutput = ContractIdInput;
// ...
const storageSlots: StorageSlot[] = [
  {
    "key": "1d63cc2495bbf5570c9a6d7f632018dc033107e7f4452405c44601bb771a4a5d",
    "value": "0000000000000000000000000000000000000000000000000000000000000000"
  },
  {
    "key": "1d63cc2495bbf5570c9a6d7f632018dc033107e7f4452405c44601bb771a4a5e",
    "value": "0000000000000000000000000000000000000000000000000000000000000000"
  },
  {
    "key": "38a1b038206f348ec947ba5770c6d713148c9d8bb0b312f47acc88777da1c09d",
    "value": "0000000000000000000000000000000000000000000000000000000000000000"
  }
];

export class AssetManagerInterface extends Interface {
  constructor() {
    super(abi);
  }

  declare functions: {
    add_asset: FunctionFragment;
    get_count_supported_assets: FunctionFragment;
    get_supported_asset: FunctionFragment;
    initialize: FunctionFragment;
    is_asset_supported: FunctionFragment;
    owner: FunctionFragment;
    remove_asset: FunctionFragment;
    renounce_ownership: FunctionFragment;
    transfer_ownership: FunctionFragment;
  };
}

export class AssetManager extends Contract {
  static readonly abi = abi;
  static readonly storageSlots = storageSlots;

  declare interface: AssetManagerInterface;
  declare functions: {
    add_asset: InvokeFunction<[asset: AssetIdInput], void>;
    get_count_supported_assets: InvokeFunction<[], BN>;
    get_supported_asset: InvokeFunction<[index: BigNumberish], Option<AssetIdOutput>>;
    initialize: InvokeFunction<[], void>;
    is_asset_supported: InvokeFunction<[asset: AssetIdInput], boolean>;
    owner: InvokeFunction<[], Option<IdentityOutput>>;
    remove_asset: InvokeFunction<[index: BigNumberish], void>;
    renounce_ownership: InvokeFunction<[], void>;
    transfer_ownership: InvokeFunction<[new_owner: IdentityInput], void>;
  };

  constructor(
    id: string | AbstractAddress,
    accountOrProvider: Account | Provider,
  ) {
    super(id, abi, accountOrProvider);
  }
}